moved functionality from pageLayout to UserNav, styled the toggle button
