+++ client/src/components/UserNav/_usernav.scss
@@ -1,4 +1,5 @@
 $userNav-z-index: 20;
+
 .userNav__wrapper {
   display: flex;
   justify-content: center;
@@ -6,6 +7,33 @@ $userNav-z-index: 20;
   position: fixed;
   width: 100%;
   z-index: $userNav-z-index;
+
+  .userNav__toggle {
+    position: absolute;
+    top: 15px;
+    left: 15px;
+    z-index: 30;
+
+    width: 50px;
+    height: 50px;
+
+    button {
+      width: 100%;
+      height: 100%;
+      
+      background-color: transparent;
+      border: none;
+      font-size: 32px;
+      
+      display: grid;
+      place-items: center;
+      border-radius: 10px;
+
+      &:hover {
+        background-color: rgba(0, 0, 0, 0.08);
+      }
+    }
+  }
 }
 
 .userNav__shadow {
@@ -68,7 +96,7 @@ $userNav-z-index: 20;
 
   .logo {
     width: 100%;
-    margin-top: 75px;
+    margin-top: 80px;
     display: flex;
     flex-direction: column;
     align-items: center;
